@@ -1,10 +1,10 @@
 module github.com/samuelyuan/TOAWMap
 
-go 1.22.0
+go 1.24.0
 
 require github.com/fogleman/gg v1.3.0
 
 require (
 	github.com/golang/freetype v0.0.0-20170609003504-e2365dfdc4a0 // indirect
-	golang.org/x/image v0.23.0 // indirect
+	golang.org/x/image v0.32.0 // indirect
 )
